Move auth to workspace crate
