+++ backend/Cargo.toml
@@ -1,9 +1,24 @@
+[workspace]
+members = ["auth"]
+resolver = "2"
+
+[workspace.dependencies]
+async-trait = "0.1.59"
+axum = { version = "0.8.6", features = ["multipart"] }
+jsonwebtoken = "9"
+serde = { version = "1.0", features = ["derive"] }
+time = "0.3"
+tracing = "0.1.41"
+uuid = { version = "1", features = ["v4", "serde"] }
+zeroize = "1.7"
+
 [package]
 name = "life-manager"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+auth = { path = "auth" }
 async-trait = "0.1.59"
 axum = { version = "0.8.6", features = ["multipart"] }
 axum-server = "0.7.3"
@@ -16,7 +31,6 @@ diesel = { version = "2.2.0", features = ["sqlite", "chrono", "uuid", "returning
 libsqlite3-sys = { version = "0.35", features = ["bundled"] }
 dotenvy = "0.15.0"
 image = "0.25.5"
-jsonwebtoken = "9"
 lazy_static = "1.4"
 ollama-rs = { version = "0.3.2", features = ["stream"] }
 once_cell = "1"
@@ -25,14 +39,12 @@ reqwest = { version = "0.12", features = ["json", "multipart", "rustls-tls"] }
 serde_json = "1.0.68"
 serde = { version = "1.0", features = ["derive"] }
 tempfile = "3"
-time = "0.3"
 tokio = { version = "1.47.1", features = ["full"] }
 tower = "0.5.2"
 tower-http = { version = "0.6.6", features = ["trace", "cors"] }
 tracing = "0.1.41"
 tracing-subscriber = { version = "0.3.20", features = ["env-filter", "fmt", "json", "time"] }
 uuid = { version = "1", features = ["v4", "serde"] }
-zeroize = "1.7"
 
 [dev-dependencies]
 axum-test = "18.1.0"
