+++ backend/auth/src/auth_state.rs
@@ -1,0 +1,6 @@
+use std::sync::Arc;
+
+use crate::auth_use_cases::AuthUseCases;
+
+#[derive(Clone)]
+pub struct AuthState(pub Arc<AuthUseCases>);
