+++ backend/auth/src/lib.rs
@@ -1,0 +1,16 @@
+mod auth_state;
+mod auth_use_cases;
+mod auth_user;
+mod jwt_secret;
+mod login_handler;
+mod login_request;
+mod login_service;
+pub mod router;
+mod test_protected_endpoint_handler;
+
+pub use auth_state::AuthState;
+pub use auth_use_cases::AuthUseCases;
+pub use auth_user::AuthUser;
+pub use login_request::{Claims, LoginRequest, LoginResponse};
+pub use login_service::{LoginResult, LoginService};
+pub use router::auth_router;
