+++ backend/auth/src/router.rs
@@ -1,0 +1,21 @@
+use axum::{
+    Router,
+    extract::FromRef,
+    routing::{get, post},
+};
+
+use crate::{
+    auth_state::AuthState,
+    login_handler::login,
+    test_protected_endpoint_handler::test_protected_endpoint,
+};
+
+pub fn auth_router<S>() -> Router<S>
+where
+    S: Clone + Send + Sync + 'static,
+    AuthState: FromRef<S>,
+{
+    Router::new()
+        .route("/login", post(login))
+        .route("/protected", get(test_protected_endpoint))
+}
