+++ backend/auth/src/test_protected_endpoint_handler.rs
@@ -1,6 +1,6 @@
 use axum::response::IntoResponse;
 
-use crate::infrastructure::auth::auth_user::AuthUser;
+use crate::auth_user::AuthUser;
 
 pub async fn test_protected_endpoint(AuthUser { user_id: user }: AuthUser) -> impl IntoResponse {
     format!("Hello {}", user.to_string())
