+++ backend/src/infrastructure.rs
@@ -1,5 +1,5 @@
 pub mod app_state;
-pub mod auth;
+pub mod auth_integration;
 pub mod db;
 pub mod document;
 pub mod document_text_extraction;
@@ -7,4 +7,5 @@ pub mod http_client;
 pub mod noop_document_text_reader;
 pub mod ollama_document_summarizer_adapter;
 pub mod reqwest_http_client;
+pub mod superuser_only_login_service;
 pub mod tesseract_adapter;
