--- backend/src/infrastructure/auth.rs
@@ -1,10 +1,0 @@
-pub mod auth_router;
-pub mod auth_state;
-pub mod auth_use_cases;
-pub mod auth_user;
-pub mod jwt_secret;
-pub mod login_handler;
-pub mod login_request;
-pub mod login_service;
-pub mod superuser_only_login_service;
-pub mod test_protected_endpoint_handler;
