+++ backend/src/infrastructure/auth_integration.rs
@@ -1,11 +1,7 @@
-use std::sync::Arc;
-
+use auth::AuthState;
 use axum::extract::FromRef;
 
-use crate::infrastructure::{app_state::AppState, auth::auth_use_cases::AuthUseCases};
-
-#[derive(Clone)]
-pub struct AuthState(pub Arc<AuthUseCases>);
+use crate::infrastructure::app_state::AppState;
 
 impl FromRef<AppState> for AuthState {
     fn from_ref(app_state: &AppState) -> Self {
