+++ backend/tests/auth_tests.rs
@@ -2,7 +2,7 @@ mod common;
 
 use crate::common::setup::{build_auth_header, run_test_with_test_profile};
 use axum_test::TestServer;
-use life_manager::infrastructure::auth::login_request::LoginRequest;
+use auth::LoginRequest;
 use reqwest::{ClientBuilder, Error, Response};
 use serial_test::serial;
 use tracing_test::traced_test;
